requirements.txt
# Core requirements
numpy>=1.24.0
colorama>=0.4.6
pyyaml>=6.0
json5>=0.9.14
prompt_toolkit>=3.0.38
rich>=13.5.0